close #2 refactor project structure
