+++ src/App.js
@@ -1,23 +1,11 @@
-import logo from './logo.svg';
-import './App.css';
+import "./styles/App.css";
 
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+    <div className="bg-gray-100 min-h-screen flex items-center justify-center">
+      <h1 className="text-4xl font-bold text-primary">
+        Welcome to Tailwind CSS
+      </h1>
     </div>
   );
 }
